[BOT] Don't cancel dependent jobs (#1475)
Change the bot's behaviour so that the coverage test remains queued
permanently. This means it will run whenever the Linux test passes.

Previously the coverage test was cancelled whenever the Linux test
failed or was cancelled, which was especially problematic in the case
of a false positive or an accidental cancellation. A complicated manual
triggering of the coverage test is now avoided.