Fix review status (#1473)
Make the bot collect all reviews instead of only the first 30 (as given by
the GitHub API by default). This is done by collecting pages with 100
reviews per page, until there are none left.

With all the reviews the bot can make a correct decision on whether
a review should be re-requested and what is the review state of a PR.

Additionally ci_tools/bot_clean_up.py is modified to avoid duplicating
code.