[BOT] Fix docs for ci_tools (#1483)
Coverage should not raise an issue if an `except` statement is not
covered (we do not enforce coverage for error statements such as `raise`
and `errors.report`, this is similar and the body of the `except` clause
will still raise an issue if anything more complex is used). This PR
adds `except` to the exceptions.

Correcting this leads to an error with the documentation. The error
arises because `numpydoc` imports all files that it tests so all
possible dependencies must be installed. This is corrected in this PR.
This issue has previously been misunderstood and fixed by trying to
import from different locations. The unnecessary code is therefore
removed.