Cleanup some datatype handling (#1525)
Remove `dtype_registry` which largely overlapped with
`dtype_and_precision_registry` (often imported and renamed to
`dtype_registry` with further confusion). Use datatypes as keys rather
than strings (which may be different depending on whether they are
obtained via `dtype.name`, `str(dtype)` or `_print(dtype)`).

In order to make this change, implement a `__reduce__` method for
DataTypes. This means that either devs should delete all .pyccel files
in their stdlib folder, or we should update the version number.

Additionally there is a minor bug fix. The LAPACK tests were raising an
import warning. It was caused by trying to import functions from the
internals via the externals. This change removes the warning by also
searching for imported objects in the imports of the imported file.