Don't auto-create checklists to fix fork problems (#1549)
As described in #1548 collaborators who are not associated with the
Pyccel organisation and contribute via forks are unable to check items
on checklists created by the bot. Here we ensure the checklist is in a
comment belonging to the person who needs to check the boxes.

Also fix new bug in welcome message.