Change default type handling behaviour (#1523)
Move the handling of type annotations from the syntactic stage to the
semantic stage to match what is described in the developer docs
[here](https://github.com/pyccel/pyccel/blob/devel/developer_docs/type_inference.md).
Fixes #1517 .

Require Python 3.8 in order to use `math.prod`.

**Commit Summary**
- Remove unnecessary body argument from `FunctionAddress` (requires a
new `__getnewargs__` implementation)
- Fix docstrings
- Add `FunctionTypeAnnotation`
- Allow devs to override the line/column in the error messages
- Rename `Type->`VariableType` for clarity
- Allow a Variable header to be annotated with a `UnionTypeStmt` for
consistency
- Duplicate the type parser (necessary as `textx` only creates the
object described in the first line of the grammar file)
- Fix the scoping of `DottedName` symbols to ensure they are not
unnecessarily renamed
- Allow symbolic aliases to be saved to a Scope. These act similarly to
`constexpr` variables in C++, we are able to use their values directly
- Visit the type annotations in the semantic stage (fixes #1518 fixes
#633 fixes #387)
- Create arguments in a `FunctionDef` by iterating over the templates
using symbolic aliases. The arguments are then created using the type
annotations. This creates n functions where n is a product of the number
of types in each union type (>=1). Handling templates via symbolic
aliases makes it easier to handle templates and union types
simultaneously. It also ensures that the template type symbol is handled
identically to the way the type it represents would be (fixes #1336)
- Pack all known type information into `AnnotatedPyccelSymbol` objects.
- Remove unused typing types
- Remove MPI deprecated class header
- Fix lapack precision typos
- Remove undocumented type syntax: `int*8`
- Remove support for header annotations combined with argument type
annotations
- Move relevant error tests to new stages
- Remove support for specifying header annotations in a separate file
(the test `tests/pyccel/test_pyccel.py::test_headers` created a header
file containing the types description and a definition file which
imported the definition).
- Create static versions of the `TypedAstNode` class methods to retrieve
type information from non-string annotations
- Allow the `fst` of a `FunctionDefArgument` to be set via the
constructor