Improve TypedAstNode to fully describe container types (#1590)
Add an attribute `class_type` to `TypedAstNode` (fixes #1530). This
attribute describes the type of the object as it would be reported in
Python. As a result it can be used to store container types (e.g.
ndarray, lists, sets, etc). Add docstrings.

Simplify type deductions using `__add__` to combine types. This function
is called repeatedly. As a simple function, it is therefore cached.