Fix compiler configuration path resolution (#1617)
Move the call to `os.chdir` to change working directory at the
last possible minute. Fixes #1614.

When a custom compiler configuration file is used Pyccel cannot change
the file name to an absolute path as we do not immediately know whether
the argument contains a path or simply a key word indicating the
compiler family. This should not cause problems, as long as the working
directory remains unchanged. As the working directory was changed to
the `__pyccel__` folder too early, we encountered the problem reported in
https://github.com/pyccel/psydac/issues/341#issuecomment-1807180073.