Fix infinite loop when slices are found in `array` call (#1615)
Remove wrong assumption that elements of a list passed to a call
to `array` are always scalars, which resulted in an infinite loop when
a slice was encountered. To this end handle a multi-dimensional
`IndexedElement` in the same way as a `Variable`. This fixes #1613.

Add a condition to ensure that an error is raised if this problem
re-arises, rather than entering an infinite loop.