Handle scalar optionals as per the standard (#1629)
Make the handling of scalar optional arguments compatible with all
compilers (fixes #1628). To this end move the creation of local versions
of the optional variables from the C-codegen stage (where it was not
correctly implemented) to the semantic parser.

Further, add tolerances in unit tests where they are missing for floating
point functions.