Fix scoping issue (#1652)
Ensure that the symbols in all enclosing scopes are checked when looking
for collisions. This fixes #1651 where the symbol associated with the
type `A` was not present in the function scope but was correctly defined
in the global scope.

Fixing this brought to light a few places in the wrapper where the
scoping was not correctly used. Notably not all variables inserted into
the scope provided their original name, which lead to unnecessary calls
to `get_expected_name`. The names of the arguments and results of the C
call are defined in a loop to improve readability and make it simpler to
match the logic used in `_wrap_BindCFunctionDefResult` where basic
scalar types inherit their logic from `_wrap_FunctionDefResult`.