Manage Pylint errors via bot (#1653)
Modify the pylint test. Pylint is now run on all files with a Python
extension. The git diff is used to filter the results to avoid an
extreme number of errors. Errors are only reported on lines that were
changed. This means new unused-imports will not be reported for now.
This mostly removes the necessity for Codacy for Python files (although
it may still be useful for some corner cases) and fixes #1648 . Errors
are not reported for cyclic-imports as they seem to be reported on lines
unrelated to the error.

Additionally some pylint errors are fixed and the ci_tools folder is
added to the PYTHONPATH for the docs test to prevent import issues
causing failing tests.