Update Intel environment (#1672)
Update the Intel environment to avoid incompatibilities due to different
versions for the C/C++ compiler and the Fortran compiler. Mark all
failing tests for Intel and link them to the corresponding new issues.
Change the default Intel compilers to the new versions from
Intel-OneApi. (Intel are deprecating `icc` and `ifort` this year.)