Add timers and fix set usage (#1711)
Add some timers around the main parts of the Pyccel command and add
a flag which allows the results to be printed. Identify the wrapper creation
as a bottleneck, partly caused by the misuse of sets in the `naming`
module. Fix this problem, resulting in a 2x speed up (~x2) in the wrapper
creation.

**Commit Summary**

- Add a `time_execution` flag to allow users to time the execution of
critical parts
- Use `a in set` syntax to benefit from set optimisations
- Only use `.lower()` when testing collisions in Fortran code
- Ensure temporaries are not renamed on insertion to scope of module
initialisation function.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>