Allow class usage in a project (#1707)
Add a header file to the wrapper. The header file has a different
behaviour depending on where it is included from. When the header
is imported from the associated wrapper, it contains the declarations
of all types and functions used in the module. When it is imported
from another file it contains a definition of a Capsule and macros
which expose the type declarations. Fixes #1650.

The implementation follows the tutorial here:
https://docs.python.org/3/extending/extending.html#using-capsules.

**Commit Summary**

- Modify the pylint flags from `-rXx` to `-rx`. Pylint was updated and
reports now show backtraces making it hard to find errors.
- Add the datatype `CStackArray` and associated class definition
`StackArrayClass` to print C-style arrays (e.g. `float a[3]`)
- Change inheritance of `PyBuildValueNode` and `PyModule_AddObject` to
inherit from `PyccelInternalFunction` and indicate that they represent
functions
- Remove `Errors.parser_scope` (superceded by
`pyccel.utilities.stage.PyccelStage`)
- Allow adding extra library dependencies to a `CompileObj`
- Allow a compiler to link libraries using absolute paths
- Expose the Python shared library suffix stored in a compiler
- Ensure that the names of functions in imported classes match those
that were printed by renaming (when stub files are introduced this
should become possible via annotations)
- Print a `ModuleHeader` for a C-Python wrapper
- Add translated modules as library dependencies for wrappers
- Wrap imports to use classes from other modules in wrappers
- Add a test involving classes in a project