Wrap class variables (#1706)
Add the wrapping of class variables. Getters and setters are created
which can be called from Python. These getters and setters are packed
into a class called `PyGetSetDefElement` to expose the property. Fixes
#1618 .

**Commit Summary**

- Update documentation to remove the reference to incomplete wrapping
- Add `BindCClassProperty` to hold the getters and setters necessary to
wrap a class property in Fortran and expose it to C
- Add `PyGetSetDefElement` to hold the getters and setters necessary to
wrap a class property in C and expose it to Python
- Update version to 1.11.0
- Remove obsolete xfail mark