List support: append method (#1709)
Add support for the append() method of Python lists to the semantic
stage. Add handling of that method to the Python printer. This fixes
#1689.

**Commit Summary**

- Add a class representation for the `append()` method, inheriting from
`PyccelInternalFunction`. The constructor ensures compatibility between
the datatypes of the `append()` argument and the list for homogenous
lists.
- Create `ClassDef` representing a `NativeHomogeneousList`. It includes
a method `append()` implemented by the `ListAppend` class.`
- Update condition in `get_cls_base()` function to properly check on the
classtype of the containers.
- Add method `_print_ListAppend()` to the Python printer for
generating the appropriate Python code for the `append()` method.
- Add check on the semantic to ensure that the correct class
definition (`PythonList`) is assigned to variables created via
expressions such as `a = [1]`
- Add tests for the `append()` method involving many scenarios where the
function could be used.

---------

Co-authored-by: EmilyBourne <louise.bourne@gmail.com>
Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>