Minor bot fix (#1733)
A recent bot fix caused a small bug. The variable `states` is used to
check if the tests which have already run have passed. However the
existing implementation doesn't filter out tests like `Bot / Thanks for
the Pull Request! (pull_request_target)` or `Bot / Update Review Stage
(pull_request_review)`. These checks are not obligatory and are
sometimes skipped so this prevents the bot from adding the labels and
marking a PR as ready.