List support/insert method (#1727)
Add support for the insert() method of Python lists to the semantic stage.
Add handling of that method to the Python printer. This fixes #1692.

**Commit Summary**

- Add a class representation for the `insert()` method, inheriting from
`PyccelInternalFunction`. The constructor ensures compatibility between
the datatypes of the `insert()` argument and the list for homogenous lists.
- Update the `_class_type` attribute for methods that return `None` 
- Add the python printer `_print_ListInsert()` to generate the
appropriate Python representation of the `insert()` method.
- Add success and error tests for the `insert()` method involving many
scenarios where the function could be used.

---------

Co-authored-by: EmilyBourne <louise.bourne@gmail.com>