Set support: constructor and `add` method (#1757)
Support the creation of Python sets of scalars with the `{}` operator.
Support the `add()` method of Python sets in the semantic stage. 
Handle that method in the Python printer. This fixes #1738.

**Commit Summary**

- Add the singleton class `NativeHomogeneousSet` to represent the
homogenous set type in `pyccel.ast.datatypes`.
- Add a `PythonSet` class to represent a call to Python's `{*args}` in
`pyccel.ast.builtins`.
- In `pyccel.parser.syntactic`, add `_visit_Set` which calls the
`PythonSet` class.
- Add `_visit_PythonSet` in `pyccel.parser.semantic` and
`_print_PythonSet` to the Python printer for generating appropriate
Python code.
- Create a file named `set_methods.py` in `pyccel/ast/builtin_methods/`.
- Add a class representing the `add()` method, inheriting from
`PyccelInternalFunction`.
- Create a `ClassDef` object representing the class
`NativeHomogeneousSet`. This class includes an `add()` method
implemented by the `SetAdd` class.
- Update the condition in the `get_cls_base()` function to properly
check the class type of the containers.
- Add the `_print_Setadd()` method to the Python printer for generating
the appropriate Python code for the `add()` method.
- Add tests for the `add()` method involving various scenarios where the
function could be used.