Improve argument singletons (#1760)
The class `ArgumentSingleton` currently lets us create singleton classes
such that there is only one instance of the class for any given set of
arguments. However our current implementation relies on `*args,
**kwargs` to guarantee compatibility with the `__init__` function of the
class. This results in `functools.signature` reporting the wrong
signature. As a result numpydoc insists on docstrings matching the
prototype of `ArgumentSingleton.__call__`. This is manageable when there
are few classes using the metaclass (as is the case currently), but it
is unhelpful when the class is used more widely (as is the case in
#1756).

The solution to this is to use the `__signature__` property
documented in `inspect.signature` to trick numpydoc into seeing the
signature of the `__init__` function.

Further, the `Singleton` metaclass does not subclass `ArgumentSingleton`
anymore. This allows for a significant simplification of the code.