Mark the failing Intel test (#1784)
This is a small PR to mark the failing Intel test `test_numpy_isfinite`,
which fails when passing `np.inf`. It seems likely that this could be
due to a difference in how the Intel and GNU compilers handle infinity.
Flagging this test as `xfail` will ensure that the other tests are still
run so other errors can be detected.