List support: extend method (#1731)
Add support for the `extend` method of Python lists to the semantic stage.
Add handling of that method to the Python printer. This fixes #1694.

**Commit Summary**
- Create a superclass `ListMethod` inheriting from
`PyccelInternalFunction`, representing the structure for all list
methods.
- Create one dynamic printer `_print_ListMethod` for all list methods to
prevent code duplication.
- Add the `ListExtend` class only to document the behavior of `extend()`
and its Pyccel implementation.
- Create `_visit_ListExtend`, which, depending on the argument passed to
`extend()`, tries to create semantic nodes of `append()` or a syntactic
node of a `For` loop and then re-parse it.
- Add tests for multiple list methods involving many scenarios where the
functions could be used.