Improve basic datatypes (#1756)
Rewrites the datatyping system and fix #1729.

As described in the issue, the existing superclass `Datatype` is
replaced with 2 superclasses representing slightly different concepts:
- `PrimitiveType` : Representing the category of datatype
(integer/floating point/etc)
- `PyccelType` : Representing the actual type of the object in Python

Subclasses of `PyccelType` generally fall into one of two categories:
- `FixedSizeType`
- `ContainerType`

These types are described in `developer_docs/type_inference.md` (it is
recommended to read this file first before reviewing this pull request.

The `precision` is removed from `ast.basic.TypedAstNode` and is stored
in the `FixedSizeType`s where it is relevant. This allows us to remove
the necessity for negative precision values as defaults, as the types
can now be easily differentiated using the different `FixedSizeType`s.
The attribute `_dtype` is removed from `ast.basic.TypedAstNode`
instances. It is still accessible via the `dtype` property but is
extracted from the class type. This ensures that the two do not deviate
from one another.

The NumPy type deduction is significantly improved due to two major
improvements. Firstly an assertion is added to `NumpyNDArrayType` to
ensure that the internal datatypes are NumPy types. This uncovered a few
bugs where the native Python types were used in our current code. Thus
the wrong type would be returned if an element of one of these objects
was returned.
Secondly `np.result_type` is now used to deduce the result type of NumPy
functions. This function is implemented by NumPy which guarantees that
it handles all corner cases and matches the expected behaviour of NumPy.
Fixes #1763.

The `CustomDatatype` is simplified slightly to remove the unused
parameters.

The printing of floats/complexes in Fortran is improved to stop printing
decimal digits beyond the precision of the number. Additionally printing
the return value of functions returning multiple results of different
types is fixed.

---------

Co-authored-by: Farouk-Echaref <farouk.echcharef@usmba.ac.ma>
Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>