Clean up unnecessary arguments in the codegen stage (#1791)
The codegen stage contains lots of old code which sometimes refers to
parameters which no longer exist. Here we clean up some of that code
and remove commented dead code which refers to it. Specifically the
functions `fcode`, `ccode` and `pycode` are removed as well as
`Codegen.doprint` and `Codegen.set_printer`. An instance of `Codegen`
must now be created with the language passed as argument.