List support: `remove` method (#1790)
Add support for the `remove()` method of Python lists to the semantic
stage and Python codegen. This fixes #1693.

**Commit Summary**

- Add a `ListRemove` subclass of `ListMethod`, representing the
`remove()` method. The constructor checks for compatibility between the
type of the `remove()` argument and the type of the list, and raises an
error if they are not homogeneous.
- Add success and error tests for the `remove()` method involving many
scenarios where the function could be used.