List support: `copy` method (#1797)
Add support for the `copy()` method of Python lists to the semantic
stage and Python codegen. This fixes #1696.

**Commit Summary**

- Add a class representation for the `copy()` method, inheriting from
`ListMethod`. The constructor copies the attributes of the list object,
to the current object that represents the resulting new copy of the
list.
- Add check for `HomogeneousListType` in `_print_IndexedElement` to
support indexing a multidimensional list.
- Add success and error tests for the `copy()` method involving many
scenarios where the function could be used.