Fix boolean cast in Fortran code (#1789)
Add missing cast when creating an array from an object with a different
datatype. Fixes #1785

In order to call the cast on the argument passed to `np.array` (which
may be a `InhomogeneousTupleType`) fix the type of the cast functions.

Modify `InhomogeneousTupleType.datatype` to return a `FixedSizeType` if
the datatypes of all elements are equivalent.

Also fix some minor bugs after #1756