Remove the FunctionDef, Imports and Classes from CodeBlocks (#1812)
Remove instances of `FunctionDef`, `Import` and `ClassDef` from
`CodeBlock`s. They were previously collected here in the syntactic
stage. Fixes https://github.com/pyccel/pyccel/issues/1768 . Use this
change to remove the argument `annotate` from `_visit_FunctionDef`.

In order to correctly group the imports, move the recognition of a `if
__name__ == '__main__'` statement to the syntactic stage. This allows
the code to be simplified. An error is now raised if multiple blocks
like this are found and the program is then handled in `_visit_Module`
in the semantic stage.