Clean infer_type function (#1837)
The function `infer_type` is critical to the `SemanticParser` but it has
several issues which are fixed in this PR. Firstly the dictionary that
is returned is modified to ensure that all the documented fields are
defined. Fixes #1842 . This removes the need for calls to `dict.get`
elsewhere in the code. Secondly the `is_target` property is removed.
This property is either set to False or is incorrectly set. It is not
used for calculations elsewhere.

A test is added for #1842. Further the tuple tests are updated so that
all objects that are supposed to be homogeneous are indexed by a
variable (this is not possible for inhomogeneous objects).