Use `PyccelAstNode.pyccel_staging` in `SemanticParser` (#1839)
Use the `pyccel_staging` property of `PyccelAstNode` objects
to facilitate the creation of partially syntactic object (e.g. when
unravelling an assignment where the `rhs` has been visited and
the `lhs` is an inhomogeneous tuple).

This is done by exiting `SemanticParser._visit` immediately if the
object passed to the function is already a semantic object.

As the property is now always used it is obvious when it has not
been correctly set. These cases are therefore fixed.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>