Use Python version to make artifact name unique (1 line) (#1899)
The updated version of `upload-artifact` means that you cannot upload
multiple artifacts with the same name. This is problematic on the
`devel` branch as the same job is run for multiple Python versions. The
easy fix for this is to add the version to the artifact name. Local
tests show that the artifacts can still be retrieved with a partial
name.