Minor coverage bot issue (4 lines) (#1907)
The coverage bot is only showing 1 line errors. This makes it hard to
understand the error as people often assume that the error is on all
displayed lines (6 lines are shown for context when the error is on a
single line). This is not what the coverage bot is designed to do but
two minor errors in the script are causing such display problems.
1. The last_valid_line_idx was incorrectly defined
2. `i` was used for a loop inside a loop over `i` leading to the value
in the outer loop being incorrectly overwritten