Fix unreleased file lock (#1931)
The C tests occasionally fail in a way that causes a deadlock which only
resolved when the code times out after an hour. This was happening due
to 2 problems:
1. `copy_internal_library` was not designed to handle copying folders
with subfolders leading to a `PermissionError` when calling `os.remove`
2. If an unhandled error is raised after the file lock is acquired the
lock is never released.

This PR fixes this problem by handling subfolders using `os.walk` and
ensuring that locks are always released using a context manager.

This change is not mentioned in the CHANGELOG as the bug was introduced
after the last release.