Clean STC imports (#1930)
Handling imports by separating concepts with limiters is potentially
error prone. It also results in lots of special cases being written to
handle each individual import (e.g. vec vs. Set_extensions). This PR
unifies the strategy by using existing Pyccel concepts. An `Import` now
takes the source file as the source argument (instead of the string
containing limiters) and an `AsName` as the target. The object in the
`AsName` is the PyccelType being handled by this import (described by a
VariableTypeAnnotation) and the `target` (local name) of the `AsName` is
the C name of this object.

**Commit Summary**
- Improve `AsName.__repr__` to help debugging when the object has no
name
- Ensure `Import` targets are ordered
- Unify the header guard strategy
- Use Pyccel objects to group concepts instead of limiters