[DICT] Add Python support for dict() function (#1897)
Add Python support for `dict()` function as a copy operator or with key
word arguments to build a dictionary. This fixes #1894 . The
introduction of the `dict` function allows variable annotations to be
supported. `__eq__` and `__hash__` functions are added to container
types. This allows the initialisation of empty containers with syntax
like `a : dict[str,int] = {}`.