[DICT] Add support for pop() method (#1891)
Add support for the `dict` method `pop()` and the initialisation of a
dictionary via a call to `{}`. Syntactic, semantic and Python printing
support is added. The `DictType` datatype is expanded to match the
expected description from the docs. The class `PythonDict` is added to
manage the initialisation. The abstract class `DictMethod` and the
subclass `DictPop` are defined to handle the `pop()` method. Fixes
#1886.

_Blocked by #1897_