Optimize substitute method (#1938)
### Description
This PR addresses an optimization for the substitute method in the
PyccelAstNode class. The current implementation does not check if
original and replacement are empty, leading to unnecessary recursion
over the code. This update adds a condition to return immediately if
original is empty (and consequenty replacement), thereby improving the
performance of the method.

---------

Co-authored-by: Emily Bourne <louise.bourne@gmail.com>