Bot improvements (#1949)
Change the key for the `check_runs` dict so that information about the
Python version is not lost. This allows results from multiple Python
versions to be considered together. This should let the coverage test
use results from multiple Python versions.

It is extremely difficult to test this kind of PR. Please review as well
as you, it will then have to be tested in the devel branch on #1906 and
minor fixes may need to be done quickly after the merge