Fix usage of used_names in pyccel_to_sympy (#1958)
`used_names` in `Scope` used to be a `set` however it was changed to a
`dict` to map the Python unique names to their low-level equivalent. In
`pyccel_to_sympy` there was 1 line which still used `used_names` as
though it were a set. This causes errors if the code manages to hit this
code block (see #1924 ). This MR corrects the docstrings which still
described `used_names` as a `set` and fixes the broken line. Fixes #1924