Remove use of ndarrays when returning arrays from Fortran functions (#1968)
Remove the use of ndarray_t in the C wrapper of a function translated to
Fortran returning arrays. Fixes
https://github.com/pyccel/pyccel/issues/1967 , making it easier to
remove ndarray_t later and reducing complexity.

The functions `_wrap_FunctionDefResult` and
`_wrap_BindCFunctionDefResult` are modified to return a dictionary with
the useful information that they calculate. In addition to the body
returned previously the function result arguments are additionally
returned. This makes it easier to pass multiple variables to the
function and simplifies `_wrap_FunctionDef` slightly.

The new `PyArray` constructing function `to_pyarray` fixes the memory
leak identified in #1180 .

**Commit Summary**
- Add an `is_alias` property to `FunctionCall` which identifies if the
returned object is a pointer.
- Fix a bug in `convert_to_literal` where booleans were converted to
integers (because `isinstance(True, int)` returns true)
- Add a new `to_pyarray` function to build a `PyArray` from a data
pointer and a stack array of shapes
- Use `var` to avoid the repetition of `expr.variable` in
`CCodePrinter_print_Declare`
- Ensure `Declare.value` is used in C code.
- Allow CStackArray objects to be declared
- Initialise local `void*` variables to `NULL` (fixes macosx warnings)
- Create `CToPythonWrapper._call_wrapped_function` to reduce code
duplication
- Use variables as keys to `FunctionDef.result_pointer_map` instead of
`FunctionDefResult` objects. This ensures that the map can also be used
in `CToPythonWrapper` when translating to Fortran
- Return a dictionary from `_wrap_FunctionDefResult` and
`_wrap_BindCFunctionDefResult` to improve readability and code locality
- Determine whether the returned array may need deallocating by Python
(this should be the case unless the result is a pointer to an argument
or is stored in a class).
- Add `capsule_cleanup` function to ensure that arrays created by Pyccel
and returned will be freed (fixes memory leak).