Remove InhomogeneousTupleVariable (#1838)
Remove the `InhomogeneousTupleVariable` class. Fixes #1583 There is now
enough information in the types to make this class unnecessary. Removing
it allows inhomogeneous tuples to be used in classes (as it was
previously not possible to combine `DottedVariable` and
`InhomogeneousTupleVariable`).

**Commit Summary**
- Use the `symbolic_alias` category in a `Scope` to map an indexed
element of an inhomogeneous tuple to the equivalent variable.
- Create `_build_X` functions for functions which accept inhomogeneous
tuple arguments (created `SemanticParser._build_PythonTupleFunction` and
`SemanticParser._build_NumpyArray`). This allows the variables to be
collected from the scope during the building process
- Allow a `PyccelUnarySub` to be used as an index (to facilitate
indexing with literals)
- Add a `is_literal_integer` function to determine if an expression is a
literal integer
- Recognise inhomogeneous tuples via their class type
- Remove `InhomogeneousTupleVariable`
- Ensure `IndexedElement` can handle inhomogeneous tuple objects.
- Don't declare inhomogeneous tuples (their elements are now already
present in the scope)
- Clarify use of dtype/class type in `FCodePrinter._print_Declare`
- Correct type annotation printing in Python (quotes are no longer
present to be removed)
- Raise a warning if trying to wrap tuples
- Clean up tuple handling in `_visit_Assign`. Fixes #1392 
- Extract printed inhomogenous tuple variables into `PythonTuple` at the
semantic stage
- Increase coverage
- Add a test for inhomogeneous tuples in classes